+++ src/s3/s3.module.ts
@@ -1,7 +1,7 @@
 import { Module, forwardRef } from '@nestjs/common';
 import { S3ClientConfig, S3Client } from '@aws-sdk/client-s3';
 import { AttendeeController } from 'src/attendees/attendees.controller';
-import { S3ModuleService } from './s3-module.service';
+import { S3ModuleService } from './s3.service';
 import { AttendeesModule } from '../attendees/attendees.module';
 
 const s3ClientConfig: S3ClientConfig = {
